bundling fixes and travis
